netops-backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "netops-backend"
version = "0.1.0"
description = "NetOps 后端服务"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.12",
    "starlette>=0.40.0,<0.47.0",
    "uvicorn>=0.30.0",
    "SQLAlchemy==2.0.40",
    "alembic==1.15.1",
    "aiosqlite==0.21.0",
    "asyncpg==0.30.0",
    "psycopg2-binary>=2.9.0",
    "redis==4.5.4",
    "python-jose==3.4.0",
    "passlib==1.7.4",
    "python-multipart==0.0.6",
    "bcrypt==4.0.1",
    "python-ldap==3.4.3",
    "ldap3==2.9.1",
    "PyJWT==2.8.0",
    "email-validator==2.2.0",
    "pyotp>=2.9.0",
    "qrcode[pil]>=7.4.0",
    "python-dotenv==1.0.1",
    "pydantic==2.10.6",
    "pydantic-settings==2.2.1",
    "httpx==0.28.1",
    "requests>=2.28.0",
    "APScheduler==3.11.0",
    "tzlocal>=3.0",
    "celery==5.3.6",
    "Pillow==10.2.0",
    "netmiko==3.4.0",
    "paramiko==3.5.1",
    "textfsm==1.1.3",
    "Jinja2==3.1.5",
    "PyYAML==6.0.2",
    "ntc-templates==7.8.0",
    "pytz==2025.1",
    "pyvmomi>=8.0.0.0",
    "boto3>=1.28.0",
    "alibabacloud_ecs20140526>=3.0.0",
    "alibabacloud_tea_openapi>=0.3.0",
]

[project.optional-dependencies]
dev = []

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
